pyproject.toml
[project]
name = "saga_bsp"
version = "0.0.1"
authors = [
  { name="Tim Noack", email="noack@esa.tu-darmstadt.de" },
]
description = "Bulk Synchronous Parallel (BSP) scheduling extension for Saga"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "scipy<1.16",
    "wfcommons==1.0",
    "anrg.saga>=0.0.11"
]
license = "MIT"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov"
]
visualization = [
    "matplotlib",
    "pygraphviz"
]
